% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.plot.R, R/pointRes-deprecated.r,
%   R/rgc.plot.R
\name{norm.plot}
\alias{norm.plot}
\alias{pointRes-deprecated}
\alias{rgc.plot}
\title{Deprecated functions in package \pkg{pointRes}.}
\usage{
norm.plot(
  list.name,
  start.yr = NULL,
  end.yr = NULL,
  sd.disp = FALSE,
  x.tick.major = 10,
  x.tick.minor = 5
)

rgc.plot(
  list.name,
  start.yr = NULL,
  end.yr = NULL,
  sd.disp = FALSE,
  x.tick.major = 10,
  x.tick.minor = 5
)
}
\description{
The functions listed below are deprecated and will be defunct in
  the near future. Help pages for deprecated functions are available at \code{help("<function>-deprecated")}.
}
\details{
pointRes-deprecated
}
\section{\code{norm.plot}}{

For \code{norm.plot}, use \code{\link{pointer.norm}}.
}

\section{\code{rgc.plot}}{

For \code{rgc.plot}, use \code{\link{pointer.rgc}}.
}

\keyword{internal}
