% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/s033.R
\docType{data}
\name{s033}
\alias{s033}
\title{Tree-ring series Schneetal}
\format{A \code{data.frame} containing 20 tree-ring series in columns and 136 years in rows.}
\usage{
data(s033)
}
\description{
This dataset presents tree-ring series for 20 European beech (\emph{Fagus sylvatica} L.) trees from the forest reserve Schneetal, Bavaria, Germany. Series are averages of two cores.
}
\references{
Principe, A.S., van der Maaten, E., van der Maaten-Theunissen, M., Jentsch, A., Wilmking, M. & Kreyling, J. (in prep.) Low resistance but high resilience in growth of a major deciduous forest tree (\emph{Fagus sylvatica} L.) in response to late spring frost in southern Germany.
}
\keyword{datasets}

